Revert "Setting up 2x2 geometry cryostat to work for FSD"
This reverts commit 1e953961556474a57d8df632d521fe13d3efb384.
These changes were made so I could use the same files for FSD and 2x2
gdml generation. While that worked, the FSD simulation had some issues
with the gdml using an assembly, but the assembly is needed for the 2x2
geometry. So instead I'm using different files to generate the FSD gdml
(see previous commit) and reverting these files to keep 2x2 geometry
generation the same.